+++ src/compute-engine/boxed-expression/boxed-symbol.ts
@@ -170,6 +170,12 @@ export class BoxedSymbol extends _BoxedExpression {
     return this.engine.box(this._id);
   }
 
+  solve(vars: string[]): null | BoxedExpression[] {
+    if (vars.length !== 1) return null;
+    if (vars.includes(this.symbol)) return [this.engine.Zero];
+    return null;
+  }
+
   get complexity(): number {
     return 7;
   }