+++ test/compute-engine/solve.test.ts
@@ -1,0 +1,77 @@
+import { engine } from '../utils';
+
+function expr(s: string) {
+  return engine.parse(s);
+}
+
+describe('expr.solve()', () => {
+  test('should solve an assignment', () => {
+    const e = expr('x = 5');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([5]);
+  });
+
+  test('should solve an assignment to a root', () => {
+    const e = expr('x = \\sqrt{5}');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([['Sqrt', 5]]);
+  });
+
+  test('should solve an assignment to a variable', () => {
+    const e = expr('x = y');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual(['y']);
+  });
+
+  test('should solve a simple equation with a variable', () => {
+    const e = expr('x - 1 + y = 0');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([['Subtract', 1, 'y']]);
+  });
+
+  test('should solve a simple equation', () => {
+    const e = expr('x + 2 = 5');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([3]);
+  });
+
+  test('should solve a quadratic equation', () => {
+    const e = expr('x^2 + 3x + 2 = 0');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([-1, -2]);
+  });
+
+  test('should solve an equation with a fractional coefficient', () => {
+    const e = expr('\\frac{2}{3}x + \\frac{1}{3} = 5');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([7]);
+  });
+
+  test('should solve an equation with a fractional root', () => {
+    const e = expr('x^2 + 2x + \\frac{1}{4} = 0');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([
+      ['Divide', ['Subtract', ['Sqrt', 3], 2], 2],
+      ['Divide', ['Subtract', -2, ['Sqrt', 3]], 2],
+    ]);
+  });
+
+  test('should solve an equation with a complex root', () => {
+    const e = expr('x^2 + 1 = 0');
+    const result = e.solve('x')?.map((x) => x.json);
+    expect(result).toEqual([['Complex', 0, 1]]);
+  });
+
+  // test('should solve an inequality', () => {
+  //   const e = expr('2x + 1 < 5');
+  //   const result = e.solve('x')?.map((x) => x.json);
+  //   expect(result).toMatchInlineSnapshot(`[]`); // @todo
+  // });
+
+  // test('should solve a system of equations', () => {
+  //   const e1 = expr('x + y = 3');
+  //   const e2 = expr('2x - y = 0');
+  //   const result = expr([e1, e2]).solve(['x', 'y']);
+  //   expect(result).toEqual(expr('x = 0, y = 3'));
+  // });
+});